Move the selection and shared helpers to nab/_run.py and the env reads to nab/env.py
