+++ pyproject.toml
@@ -62,6 +62,7 @@ exclude = [
     "tests/test_cli_docs.py",
     "tests/test_contributing_docs.py",
     "tests/test_embedding_docs.py",
+    "tests/test_env_layer.py",
     "tests/test_release.py",
     "tests/test_type_check_scope.py",
     "tests/test_workspace_shape.py",
