+++ src/nab/_config_cmd.py
@@ -26,6 +26,12 @@ from nab_project.config_sources import (
     render_list,
 )
 
+from ._run import (
+    _cli_overrides,
+    _fail_config,
+    effective_config,
+    require_pyproject_file,
+)
 from .cli import (
     BuildPolicyFlag,
     DecisionOrderFlag,
@@ -34,13 +40,9 @@ from .cli import (
     ModeFlag,
     OfflineFlag,
     ResolutionFlag,
-    _cli_overrides,
-    _fail_config,
     app,
-    effective_config,
-    printer,
-    require_pyproject_file,
 )
+from .output import printer
 
 ActionArg = Annotated[str, tyro.conf.Positional]
 KeyArg = Annotated[str, tyro.conf.Positional]
