+++ src/nab/_download.py
@@ -5,9 +5,9 @@ archive into a local directory: the union of every target's
 artefacts, deduplicated by URL, which for a declared matrix
 pre-populates a directory for offline deployment across platforms.
 
-The helpers this shares with :mod:`nab._lock` (config loading, the
-printer, transport selection, the resolve step) live in :mod:`nab.cli`;
-everything else is imported from the module that defines it.
+The helpers this shares with :mod:`nab._lock` live in :mod:`nab._run`,
+and the run's printer in :mod:`nab.output`; everything else is imported
+from the module that defines it.
 """
 
 from __future__ import annotations
@@ -21,8 +21,20 @@ import tyro
 from nab_project.download import DownloadError, download_lock
 from nab_project.resolve import build_lock_input
 
-from . import cli as _cli
-from ._lock import resolve_extra_selection, resolve_group_selection
+from ._run import (
+    _cli_overrides,
+    _layered_run_settings_or_exit,
+    _load_config,
+    _make_transport,
+    _project_cli_overrides_or_exit,
+    _reject_python_override_in_universal,
+    _resolve,
+    _resolve_effective_cache_dir,
+    project_config_overrides,
+    read_config_ladder,
+    resolve_extra_selection,
+    resolve_group_selection,
+)
 from .cli import (
     BuildPolicyFlag,
     DecisionOrderFlag,
@@ -34,7 +46,7 @@ from .cli import (
     ResolutionFlag,
     app,
 )
-from .output import ProgressReporter
+from .output import ProgressReporter, printer
 
 
 @app.command
@@ -93,7 +105,7 @@ def download(  # noqa: PLR0913 - tyro maps each kwarg to a CLI flag so a config 
         path, extras=extras, all_extras=all_extras
     )
 
-    overrides = _cli._cli_overrides(  # noqa: SLF001
+    overrides = _cli_overrides(
         cli_resolution=project_resolution,
         cli_offline=offline,
         cli_cache_dir=cache_dir,
@@ -111,23 +123,19 @@ def download(  # noqa: PLR0913 - tyro maps each kwarg to a CLI flag so a config 
         cli_base_group=project_base_group,
         cli_build_group=project_build_group,
     )
-    project_overrides = _cli.project_config_overrides(overrides)
-    _cli._project_cli_overrides_or_exit(project_overrides)  # noqa: SLF001
-    config = _cli._load_config(  # noqa: SLF001
+    project_overrides = project_config_overrides(overrides)
+    _project_cli_overrides_or_exit(project_overrides)
+    config = _load_config(
         path,
         discover_workspace=workspace_discovery,
         cli_overrides=project_overrides,
     )
-    ladder = _cli.read_config_ladder(path, overrides)
-    settings = _cli._layered_run_settings_or_exit(  # noqa: SLF001
-        ladder, produces_lock=False
-    )
-    effective_cache_dir = _cli._resolve_effective_cache_dir(  # noqa: SLF001
-        settings.cache_dir, cache=cache
-    )
-    _cli._reject_python_override_in_universal(config, python)  # noqa: SLF001
-    transport = _cli._make_transport(settings.http_backend)  # noqa: SLF001
-    result = _cli._resolve(  # noqa: SLF001
+    ladder = read_config_ladder(path, overrides)
+    settings = _layered_run_settings_or_exit(ladder, produces_lock=False)
+    effective_cache_dir = _resolve_effective_cache_dir(settings.cache_dir, cache=cache)
+    _reject_python_override_in_universal(config, python)
+    transport = _make_transport(settings.http_backend)
+    result = _resolve(
         path,
         config=config,
         cache_dir=effective_cache_dir,
@@ -138,7 +146,7 @@ def download(  # noqa: PLR0913 - tyro maps each kwarg to a CLI flag so a config 
         groups=selected_groups,
         extras=selected_extras,
         resolution_strategy=settings.resolution,
-        progress=ProgressReporter(_cli.printer()),
+        progress=ProgressReporter(printer()),
     )
     lock_input = build_lock_input(
         result,
@@ -147,7 +155,7 @@ def download(  # noqa: PLR0913 - tyro maps each kwarg to a CLI flag so a config 
         dependency_groups=selected_groups,
     )
 
-    download_transport = _cli._make_transport(settings.http_backend)  # noqa: SLF001
+    download_transport = _make_transport(settings.http_backend)
     try:
         outcome = download_lock(
             lock_input,
@@ -157,13 +165,13 @@ def download(  # noqa: PLR0913 - tyro maps each kwarg to a CLI flag so a config 
             offline=settings.offline,
         )
     except DownloadError as e:
-        _cli.printer().error(f"download failed: {e}")
+        printer().error(f"download failed: {e}")
         sys.exit(1)
     except OSError as e:
-        _cli.printer().error(f"cannot write to output directory {output}: {e}")
+        printer().error(f"cannot write to output directory {output}: {e}")
         sys.exit(1)
 
-    _cli.printer().done(
+    printer().done(
         f"Downloaded {len(outcome.written)} files,"
         f" {len(outcome.skipped)} already present, into {output}"
     )
